+++ lib/membrane/core/pipeline.ex
@@ -84,14 +84,8 @@ defmodule Membrane.Core.Pipeline do
   end
 
   @impl GenServer
-  def handle_info(Message.new(:link_response, link_id), state) do
-    state = ChildLifeController.handle_link_response(link_id, state)
-    {:noreply, state}
-  end
-
-  @impl GenServer
-  def handle_info(Message.new(:spec_linking_timeout, spec_ref), state) do
-    state = ChildLifeController.handle_spec_timeout(spec_ref, state)
+  def handle_info(Message.new(:link_response, [link_id, direction]), state) do
+    state = ChildLifeController.handle_link_response(link_id, direction, state)
     {:noreply, state}
   end