+++ lib/membrane/remote_controlled/pipeline.ex
@@ -1,7 +1,11 @@
 defmodule Membrane.RemoteControlled.Pipeline do
   @moduledoc """
-  `Membrane.RemoteControlled.Pipeline` is a basic `Membrane.Pipeline` implementation that can be
-  controlled by a controlling process.
+  Basic `Membrane.Pipeline` implementation that can be controlled by a controlling process.
+
+  The easiest way to start this pipeline is to use `start_link!/1`
+  ```
+    pipeline = Membrane.RemoteControlled.Pipeline.start_link!()
+  ```
 
   The controlling process can request the execution of arbitrary
   valid `Membrane.Pipeline.Action`:
@@ -16,24 +20,24 @@ defmodule Membrane.RemoteControlled.Pipeline do
   sent by the pipeline and later on synchronously await for these messages:
   ```
   # subscribes to message which is sent when the pipeline enters `playing`
-  Pipeline.subscribe(pipeline, %Message.Playing{})
+  Membrane.RemoteControlled.Pipeline.subscribe(pipeline, %Message.Playing{})
   ...
   # awaits for the message sent when the pipeline enters :playing playback
-  Pipeline.await_playing(pipeline)
+  Membrane.RemoteControlled.Pipeline.await_playing(pipeline)
   ...
   ```
 
   `Membrane.RemoteControlled.Pipeline` can be used when there is no need for introducing a custom
   logic in the `Membrane.Pipeline` callbacks implementation. An example of usage could be running a
   pipeline from the elixir script. `Membrane.RemoteControlled.Pipeline` sends the following messages:
-  * `Membrane.RemoteControlled.Message.Playing.t()` sent when pipeline enters `playing` playback,
-  * `Membrane.RemoteControlled.Message.StartOfStream.t()` sent
+  * `t:Membrane.RemoteControlled.Message.Playing.t/0` sent when pipeline enters `playing` playback,
+  * `t:Membrane.RemoteControlled.Message.StartOfStream.t/0` sent
   when one of direct pipeline children informs the pipeline about start of a stream,
-  * `Membrane.RemoteControlled.Message.EndOfStream.t()` sent
+  * `t:Membrane.RemoteControlled.Message.EndOfStream.t/0` sent
   when one of direct pipeline children informs the pipeline about end of a stream,
-  * `Membrane.RemoteControlled.Message.Notification.t()` sent when pipeline
+  * `t:Membrane.RemoteControlled.Message.Notification.t/0` sent when pipeline
   receives notification from one of its children,
-  * `Membrane.RemoteControlled.Message.Terminated.t()` sent when the pipeline gracefully terminates.
+  * `t:Membrane.RemoteControlled.Message.Terminated.t/0` sent when the pipeline gracefully terminates.
   """
 
   use Membrane.Pipeline
@@ -58,7 +62,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
 
   @doc """
   Starts the `Membrane.RemoteControlled.Pipeline` and links it to the current process. The process
-  that makes the call to the `start_link/1` automatically become the controller process.
+  that makes the call to the `start_link/1` automatically becomes the controller process.
   """
   @spec start_link([Pipeline.config_entry() | {:controller_pid, pid()}]) :: Pipeline.on_start()
   def start_link(options \\ []) do
@@ -66,13 +70,25 @@ defmodule Membrane.RemoteControlled.Pipeline do
     Pipeline.start_link(__MODULE__, %{controller_pid: controller_pid}, config)
   end
 
+  @spec start_link!([Pipeline.config_entry() | {:controller_pid, pid()}]) :: pid
+  def start_link!(options \\ []) do
+    {:ok, _supervisor, pipeline} = start_link(options)
+    pipeline
+  end
+
   @doc """
-  Does the same as the `start_link/1` but starts the process outside of the supervision tree.
+  Does the same as the `start_link/1` but starts the process outside of the current supervision tree.
   """
   @spec start([Pipeline.config_entry() | {:controller_pid, pid()}]) :: Pipeline.on_start()
   def start(options \\ []) do
     {controller_pid, config} = Keyword.pop(options, :controller_pid, self())
     Pipeline.start(__MODULE__, %{controller_pid: controller_pid}, config)
+  end
+
+  @spec start!([Pipeline.config_entry() | {:controller_pid, pid()}]) :: pid
+  def start!(options \\ []) do
+    {:ok, _supervisor, pipeline} = start(options)
+    pipeline
   end
 
   defmacrop pin_leaf_nodes(ast) do
@@ -107,7 +123,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
   end
 
   @doc """
-  Awaits for the first `Membrane.RemoteControlled.Message()` wrapping the `Membrane.RemoteControlled.Message.Playing()`
+  Awaits for the first `t:Membrane.RemoteControlled.Message.t/0` wrapping the `t:Membrane.RemoteControlled.Message.Playing.t/0`
   It is required to firstly use the `subscribe/2` to subscribe to a given message before awaiting
   for that message.
 
@@ -123,7 +139,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
   end
 
   @doc """
-  Awaits for the first `Membrane.RemoteControlled.Message()` wrapping the `Membrane.RemoteControlled.Message.StartOfStream()` message
+  Awaits for the first `t:Membrane.RemoteControlled.Message.t/0` wrapping the `t:Membrane.RemoteControlled.Message.StartOfStream.t/0` message
   with no further constraints, sent by the process with `pipeline` pid.
   It is required to firstly use the `subscribe/2` to subscribe to a given message before awaiting
   for that message.
@@ -140,7 +156,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
   end
 
   @doc """
-  Awaits for the first `Membrane.RemoteControlled.Message()` wrapping the `Membrane.RemoteControlled.Message.StartOfStream()` message
+  Awaits for the first `t:Membrane.RemoteControlled.Message.t/0` wrapping the `t:Membrane.RemoteControlled.Message.StartOfStream.t/0` message
   concerning the given `element`, sent by the process with `pipeline` pid.
   It is required to firstly use the `subscribe/2` to subscribe to a given message before awaiting
   for that message.
@@ -158,7 +174,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
   end
 
   @doc """
-  Awaits for the first `Membrane.RemoteControlled.Message()` wrapping the `Membrane.RemoteControlled.Message.StartOfStream()` message
+  Awaits for the first `t:Membrane.RemoteControlled.Message.t/0` wrapping the `t:Membrane.RemoteControlled.Message.StartOfStream.t/0` message
   concerning the given `element` and the `pad`, sent by the process with `pipeline` pid.
   It is required to firstly use the `subscribe/2` to subscribe to a given message before awaiting
   for that message.
@@ -176,7 +192,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
   end
 
   @doc """
-  Awaits for the first `Membrane.RemoteControlled.Message()` wrapping the `Membrane.RemoteControlled.Message.EndOfStream()` message
+  Awaits for the first `t:Membrane.RemoteControlled.Message.t/0` wrapping the `t:Membrane.RemoteControlled.Message.EndOfStream.t/0` message
   with no further constraints, sent by the process with `pipeline` pid.
   It is required to firstly use the `subscribe/2` to subscribe to a given message before awaiting
   for that message.
@@ -193,7 +209,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
   end
 
   @doc """
-  Awaits for the first `Membrane.RemoteControlled.Message()` wrapping the `Membrane.RemoteControlled.Message.EndOfStream()` message
+  Awaits for the first `t:Membrane.RemoteControlled.Message.t/0` wrapping the `t:Membrane.RemoteControlled.Message.EndOfStream.t/0` message
   concerning the given `element`, sent by the process with `pipeline` pid.
   It is required to firstly use the `subscribe/2` to subscribe to a given message before awaiting
   for that message.
@@ -211,7 +227,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
   end
 
   @doc """
-  Awaits for the first `Membrane.RemoteControlled.Message()` wrapping the `Membrane.RemoteControlled.Message.EndOfStream()` message
+  Awaits for the first `t:Membrane.RemoteControlled.Message.t/0` wrapping the `t:Membrane.RemoteControlled.Message.EndOfStream.t/0` message
   concerning the given `element` and the `pad`, sent by the process with `pipeline` pid.
   It is required to firstly use the `subscribe/2` to subscribe to a given message before awaiting
   for that message.
@@ -229,7 +245,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
   end
 
   @doc """
-  Awaits for the first `Membrane.RemoteControlled.Message()` wrapping the `Membrane.RemoteControlled.Message.Notification()`
+  Awaits for the first `t:Membrane.RemoteControlled.Message.t/0` wrapping the `t:Membrane.RemoteControlled.Message.Notification.t/0`
   message with no further constraints, sent by the process with `pipeline` pid.
   It is required to firstly use the `subscribe/2` to subscribe to a given message before awaiting
   for that message.
@@ -246,7 +262,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
   end
 
   @doc """
-  Awaits for the first `Membrane.RemoteControlled.Message()` wrapping the `Membrane.RemoteControlled.Message.Notification()` message
+  Awaits for the first `t:Membrane.RemoteControlled.Message.t/0` wrapping the `t:Membrane.RemoteControlled.Message.Notification.t/0` message
   concerning the given `element`, sent by the process with `pipeline` pid.
   It is required to firstly use the `subscribe/2` to subscribe to a given message before awaiting
   for that message.
@@ -264,7 +280,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
   end
 
   @doc """
-  Awaits for the `Membrane.RemoteControlled.Message()` wrapping the `Membrane.RemoteControlled.Message.Terminated` message,
+  Awaits for the `t:Membrane.RemoteControlled.Message.t/0` wrapping the `Membrane.RemoteControlled.Message.Terminated` message,
   which is send when the pipeline gracefully terminates.
   It is required to firstly use the `subscribe/2` to subscribe to a given message before awaiting
   for that message.
@@ -282,7 +298,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
 
   @doc """
   Subscribes to a given `subscription_pattern`. The `subscription_pattern` should describe some subset
-  of elements of `Membrane.RemoteControlled.Pipeline.message_t()` type. The `subscription_pattern`
+  of elements of `t:Membrane.RemoteControlled.Pipeline.Message.t/0` type. The `subscription_pattern`
   must be a match pattern.
 
 
@@ -310,7 +326,7 @@ defmodule Membrane.RemoteControlled.Pipeline do
   end
 
   @doc """
-  Sends a list of `Pipeline.Action.t()` to the given `Membrane.RemoteControlled.Pipeline` for execution.
+  Sends a list of `t:Pipeline.Action.t/0` to the given `Membrane.RemoteControlled.Pipeline` for execution.
 
   Usage example:
     1) making the `Membrane.RemoteControlled.Pipeline` start the `Membrane.ChildrenSpec`