+++ lib/membrane/testing/sink.ex
@@ -37,7 +37,7 @@ defmodule Membrane.Testing.Sink do
     accepted_format: _any
 
   def_options autodemand: [
-                type: :boolean,
+                spec: boolean(),
                 default: true,
                 description: """
                 If true element will automatically make demands.