+++ test/support/bin/test_bins.ex
@@ -68,8 +68,8 @@ defmodule Membrane.Support.Bin.TestBins do
     @moduledoc false
     use Membrane.Bin
 
-    def_options filter1: [type: :atom],
-                filter2: [type: :atom]
+    def_options filter1: [spec: module()],
+                filter2: [spec: module()]
 
     def_input_pad :input, accepted_format: _any, demand_unit: :buffers
 
@@ -139,8 +139,8 @@ defmodule Membrane.Support.Bin.TestBins do
     @moduledoc false
     use Membrane.Bin
 
-    def_options filter1: [type: :atom],
-                filter2: [type: :atom]
+    def_options filter1: [spec: module()],
+                filter2: [spec: module()]
 
     def_input_pad :input, demand_unit: :buffers, accepted_format: _any, availability: :on_request
 
@@ -182,8 +182,8 @@ defmodule Membrane.Support.Bin.TestBins do
     @moduledoc false
     use Membrane.Bin
 
-    def_options filter: [type: :atom],
-                sink: [type: :atom]
+    def_options filter: [spec: module()],
+                sink: [spec: module()]
 
     def_input_pad :input, demand_unit: :buffers, accepted_format: _any
 
@@ -216,8 +216,8 @@ defmodule Membrane.Support.Bin.TestBins do
     @moduledoc false
     use Membrane.Bin
 
-    def_options source: [type: :atom],
-                sink: [type: :atom]
+    def_options source: [spec: module()],
+                sink: [spec: module()]
 
     @impl true
     def handle_init(_ctx, opts) do