+++ test/support/child_removal_test/filter.ex
@@ -19,11 +19,10 @@ defmodule Membrane.Support.ChildRemovalTest.Filter do
   def_input_pad :input2, demand_unit: :buffers, accepted_format: _any, availability: :on_request
 
   def_options demand_generator: [
-                type: :function,
                 spec: (pos_integer -> non_neg_integer),
                 default: &__MODULE__.default_demand_generator/1
               ],
-              playing_delay: [type: :integer, default: 0]
+              playing_delay: [spec: integer(), default: 0]
 
   @impl true
   def handle_init(_ctx, opts) do